Move common functions to test_base.py
